% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_Clustering.R
\name{addManualClustering}
\alias{addManualClustering}
\title{Add Manual Clustering to CYTdata}
\usage{
addManualClustering(
  CYTdata,
  name,
  clusteringIDs,
  additionalPalette = NULL,
  checkOverwrite = TRUE
)
}
\arguments{
\item{CYTdata}{An object of class \code{CYTdata}. It should contain clustering data in the \code{cellClustering} slot.}

\item{name}{A character string specifying the name of the new clustering to be added. This will be used as the column name in the \code{cellClustering} slot.}

\item{clusteringIDs}{A vector of clustering identifiers (IDs) to assign to cells. The length of this vector should match the number of cells in the \code{CYTdata} object.}

\item{additionalPalette}{(optional) A vector specifying custom colors for the new clustering. The number of colors should match the number of unique clustering IDs. If \code{NULL}, the default color palette will be applied using \code{rainbow}.}

\item{checkOverwrite}{A logical value indicating whether to prompt the user for confirmation if a clustering with the same name already exists in the \code{cellClustering} slot. Default is \code{TRUE}.}
}
\value{
An updated \code{CYTdata} object with the new clustering added to the \code{cellClustering} slot. The associated color palette for the clustering is also stored in the \code{clusteringPalette} list.
}
\description{
This function allows the user to manually add a new clustering to the CYTdata object. The new clustering is stored in the \code{cellClustering} slot, and an optional color palette can be provided for visualization. The function also checks if the specified clustering name already exists and prompts the user for confirmation if overwrite is allowed.
}
\details{
The function checks if the specified clustering name already exists in the \code{cellClustering} slot of the \code{CYTdata} object. If the name exists and \code{checkOverwrite} is \code{TRUE}, the user will be asked if they want to overwrite the existing clustering. If the user confirms, the existing clustering will be replaced with the new one.

If no custom palette is provided, the function will generate a default color palette using the \code{rainbow} function.
}
\examples{
# Example usage
CYTdata <- addManualClustering(CYTdata = CYTdata,
                                name = "newClustering",
                                clusteringIDs = c(1, 2, 1, 3, 2, 1),
                                checkOverwrite = TRUE)

# Using custom color palette
customPalette <- c("red", "green", "blue")
CYTdata <- addManualClustering(CYTdata = CYTdata,
                                name = "customClustering",
                                clusteringIDs = c(1, 2, 1, 3, 2, 1),
                                additionalPalette = customPalette,
                                checkOverwrite = TRUE)

}
\seealso{
\code{\link{checkValidity}}, \code{\link{qassert}}
}
