% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_DimRed.R
\name{addManualDimRed}
\alias{addManualDimRed}
\title{Add Manual Dimensionality Reduction}
\usage{
addManualDimRed(CYTdata, DimRedName, DimRedDF, checkOverwrite = TRUE)
}
\arguments{
\item{CYTdata}{An object of class \code{CYTdata}. It should contain dimensionality reduction data in the \code{DimRedData} slot.}

\item{DimRedName}{A character string specifying the name for the new dimensionality reduction. This will be used as the key to store the dimensionality reduction in the `cellDimRed` slot.}

\item{DimRedDF}{A data frame containing the result of the dimensionality reduction. It should have rows corresponding to cells and columns representing reduced dimensions.}

\item{checkOverwrite}{A logical value indicating whether to check for the existence of the `DimRedName` and prompt the user for confirmation before overwriting an existing entry. Default is \code{TRUE}.}
}
\value{
The updated \code{CYTdata} object with the new dimensionality reduction added.
}
\description{
This function allows users to add a manually computed dimensionality reduction (DimRed) to the CYTdata object. The provided `DimRedDF` (data frame) is stored in the `cellDimRed` slot under the specified `DimRedName`. This function ensures that if the given `DimRedName` already exists, the user is prompted for confirmation before overwriting.
}
\details{
This function is useful for manually adding a result of a dimensionality reduction (e.g., PCA, t-SNE, UMAP) into the `CYTdata` object. If the specified `DimRedName` already exists in the `cellDimRed` slot, the user will be asked whether to overwrite it. This ensures that important data is not unintentionally lost.
}
\examples{
# Example usage:
CYTdata <- addManualDimRed(CYTdata = CYTdata,
                           DimRedName = "PCA_Results",
                           DimRedDF = pca_result_df)

}
