% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_CYTdata_classes.R
\docType{class}
\name{cellData-class}
\alias{cellData-class}
\alias{cellData}
\title{cellData class}
\description{
This class stores cell expression data, dimensionality reduction results, and any additional expression-related data for cells.
}
\section{Slots}{

\describe{
\item{\code{cellExprs}}{A data frame containing the expression data of cells.}

\item{\code{cellDimRed}}{A list containing the results of dimensionality reduction applied to cell expression data.}

\item{\code{cellAdditionalexprs}}{A data frame containing any additional expression-related data for cells.}
}}

