% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_Files.R
\name{checkFCS_getChannels}
\alias{checkFCS_getChannels}
\title{Check if files have the correct FCS extension, and extract and visualize the channels of FCS files}
\usage{
checkFCS_getChannels(files, metadata)
}
\arguments{
\item{files}{A character vector containing the paths to the FCS files. Each file should be in FCS format.}

\item{metadata}{A data frame containing metadata to annotate the heatmap.}
}
\value{
A list containing:
\item{tab}{A matrix of 0s and 1s indicating the presence (1) or absence (0) of each channel in each file.}
\item{Heatmap}{A heatmap visualizing the presence/absence of channels across files. The heatmap uses grey shades to indicate presence and absence of channels.}
\item{completeChannels}{A character vector containing the channels that are present in all files.}
\item{partialChannels}{A character vector containing the channels that are not present in all files.}
}
\description{
This function :
- checks whether the provided files have the `.fcs` extension. If any files do not have the correct extension, the error message displayed them.
- extracts the marker names (channels) from multiple FCS files, compares them across all files, and generates a heatmap to visualize the presence/absence across the files
}
\examples{
# Example with FCS files and metadata
files <- c("data/sample1.fcs", "data/sample2.fcs")
metadata <- data.frame(Condition = c("A", "B"))
rownames(metadata) <- gsub(pattern = ".fcs", replacement = "", x = basename(files))
result <- checkFCS_getChannels(files, metadata)

# Accessing the results
result$tab  # Presence/absence matrix
result$completeChannels  # Channels present in all files
result$partialChannels  # Channels present in some but not all files
# Example with invalid files
files <- c("data/sample1.csv", "data/sample2.fcs")
result <- checkFCS_getChannels(files, metadata)  # This will throw an error for "sample1.csv"
}
