% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_Clustering.R
\name{checkgetMetaclusteringMembership}
\alias{checkgetMetaclusteringMembership}
\title{Check Metaclustering Membership for Clusters}
\usage{
checkgetMetaclusteringMembership(CYTdata, clustering, metaclustering)
}
\arguments{
\item{CYTdata}{A `CYTdata` object containing the clustering and metaclustering data.}

\item{clustering}{A string specifying the name of the clustering column in the `cellClustering` data frame.}

\item{metaclustering}{A string specifying the name of the metaclustering column in the `cellClustering` data frame.}
}
\value{
A data frame with two columns:
- `metaclusters`: The metacluster labels corresponding to each cluster.
- `clusters`: The cluster labels.
}
\description{
This function checks the membership of clusters in a given metaclustering in the `CYTdata` object. It ensures that each cluster has cells that belong to only one metacluster and returns a data frame showing the metacluster membership for each cluster.
}
\details{
This function checks if each cluster from the specified `clustering` column belongs exclusively to one metacluster from the `metaclustering` column. If any cluster contains cells from multiple metaclusters, an error is raised. The function then returns a data frame mapping each cluster to its corresponding metacluster.

If the input `clustering` and `metaclustering` columns have multiple metacluster memberships for a cluster, the function stops and reports which clusters have cells assigned to distinct metaclusters.
}
\examples{
# Example: Get metaclustering membership for clusters in CYTdata
membership = checkgetMetaclusteringMembership(CYTdata, clustering = "clusterColumn", metaclustering = "metaClusterColumn")

}
\seealso{
\code{\link{table}} for generating contingency tables, which is used in this function to check memberships.
}
