% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_CYTdata_classes.R
\name{checkValidity}
\alias{checkValidity}
\title{Validate CYTdata Object}
\usage{
checkValidity(object, mode = c("error", "warning"))
}
\arguments{
\item{object}{A CYTdata object to be validated.}

\item{mode}{A string specifying the mode of the validation: "error" (raises errors) or "warning" (shows warnings).}
}
\value{
The input CYTdata object if all checks pass. If any validation fails, an error or warning is raised based on the mode.
}
\description{
This function validates the structure and contents of a CYTdata object, ensuring that the object and its slots are of the correct classes, and that the data within each slot meets specific criteria.
The function checks:
- Class consistency for each slot.
- Integrity of data in specific data frames and lists (e.g., no missing values, proper naming conventions).
}
