% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_CYTdata_classes.R
\docType{class}
\name{clusteringData-class}
\alias{clusteringData-class}
\alias{clusteringData}
\title{clusteringData class}
\description{
This class stores the clustering information for the cells, including cell clustering assignments and associated color palettes.
}
\section{Slots}{

\describe{
\item{\code{cellClustering}}{A data frame containing the clustering assignments for the cells.}

\item{\code{clusteringPalette}}{A list containing the color palettes used for clustering visualization.}
}}

