% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_CYTdata_classes.R
\docType{class}
\name{CYTdata-class}
\alias{CYTdata-class}
\alias{CYTdata}
\title{CYTdata class}
\arguments{
\item{cellData}{A `cellData` object containing cell-level expression and additional data.}

\item{sampleData}{A `sampleData` object containing sample-level data and metadata.}

\item{clusteringData}{A `clusteringData` object containing clustering results and associated color palettes.}

\item{differentialTesting}{A `differentialTesting` object containing results from differential testing.}

\item{Kinetic}{A `Kinetic` object containing kinetic data.}

\item{TrajectoryInference}{A `TrajectoryInference` object containing trajectory inference results.}
}
\description{
This is the main data container for all the biological data, including cell data, sample data, clustering data, and other analysis results.
}
