% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_CYTdata_classes.R
\docType{class}
\name{differentialTesting-class}
\alias{differentialTesting-class}
\alias{differentialTesting}
\title{differentialTesting class}
\description{
This class stores data related to differential testing results, typically in the form of a data frame containing statistical results for comparisons between different conditions or groups.
}
\section{Slots}{

\describe{
\item{\code{data}}{A data frame containing the differential testing results, such as p-values, test statistics, fold changes, etc.}
}}

