% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_Sampling.R
\name{downsampleData}
\alias{downsampleData}
\title{Downsample Data in CYTdata Object}
\usage{
downsampleData(
  CYTdata,
  type = c("absolute", "percent", "smallSample", "percentSample", "absoluteSample"),
  absolute = 1000,
  percentage = 1/5,
  seed = 42
)
}
\arguments{
\item{CYTdata}{A CYTdata object containing the dataset to be downsampled.}

\item{type}{A character string specifying the downsampling type. Valid options are:
- `"absolute"`: Downsample the entire dataset to a fixed number of cells.
- `"percent"`: Downsample the entire dataset by a percentage.
- `"smallSample"`: Downsample each sample to the size of the smallest sample.
- `"percentSample"`: Downsample each sample independently by a given percentage.
- `"absoluteSample"`: Downsample each sample to a fixed number of cells.}

\item{absolute}{A positive integer specifying the number of cells for downsampling (used in `"absolute"`, `"absoluteSample"`).}

\item{percentage}{A numeric value between 0 and 1 representing the proportion of the total number of cells to retain (used in `"percent"` and `"percentSample"`).}

\item{seed}{A numeric value for the random seed, to ensure reproducibility of the downsampling process.}
}
\value{
A CYTdata object containing the downsampled data.
}
\description{
This function performs downsampling on a CYTdata object. The downsampling can be done in various ways:
- **absolute**: Downsampling the entire dataset to a fixed number of cells.
- **percent**: Downsampling the entire dataset by a given percentage of cells.
- **smallSample**: Downsampling each sample to the size of the smallest sample.
- **percentSample**: Downsampling each sample independently by a given percentage.
- **absoluteSample**: Downsampling each sample to an absolute number of cells.
}
\details{
This function provides several methods for downsampling the data in the `CYTdata` object:
- If `type = "absolute"`, it will select a fixed number of cells from the entire dataset.
- If `type = "percent"`, it will select a fixed percentage of cells from the entire dataset.
- If `type = "smallSample"`, each sample will be downsampled to the size of the smallest sample.
- If `type = "percentSample"`, each sample will be downsampled by the specified percentage.
- If `type = "absoluteSample"`, each sample will be downsampled to the specified absolute number of cells.

The function updates the following components of the CYTdata object:
- `cellExprs`
- `cellAdditionalexprs`
- `cellSample`
- `sampleMetadata`
- `cellClustering`
- `cellDimRed`
}
\examples{
# Example 1: Downsample the dataset to 50\% of cells
CYTdata_downsampled <- downsampleData(CYTdata = CYTdata, type = "percent", percentage = 0.5)

# Example 2: Downsample each sample to 1000 cells
CYTdata_downsampled <- downsampleData(CYTdata = CYTdata, type = "absoluteSample", absolute = 1000)

# Example 3: Downsample the dataset by 30\% of total cells
CYTdata_downsampled <- downsampleData(CYTdata = CYTdata, type = "percent", percentage = 0.3)

# Example 4: Downsample each sample to the smallest sample size
CYTdata_downsampled <- downsampleData(CYTdata = CYTdata, type = "smallSample")

}
\seealso{
\code{\link{checkValidity}} for validating the CYTdata object before downsampling.
}
