% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_Samples.R
\name{importsampleMetadata}
\alias{importsampleMetadata}
\title{Import Sample Metadata into CYTdata Object}
\usage{
importsampleMetadata(
  CYTdata,
  sampleMetadata_df,
  optionnalPalette = NULL,
  checkOverwrite = TRUE
)
}
\arguments{
\item{CYTdata}{A CYTdata object that will be updated with the new sample metadata.}

\item{sampleMetadata_df}{A dataframe containing sample metadata. The columns of this dataframe should represent metadata variables associated with the samples.}

\item{optionnalPalette}{An optional list providing custom palettes for metadata visualization. If `NULL`, a default palette using `rainbow` is applied to each metadata variable.}

\item{checkOverwrite}{A logical value indicating whether to prompt the user to confirm overwriting existing sample metadata in the CYTdata object if it already exists. Default is `TRUE`.}
}
\value{
A modified CYTdata object with the updated sample metadata and associated metadata palette.
}
\description{
This function allows users to import a sample metadata dataframe (`sampleMetadata_df`) into a CYTdata object.
It also provides the option to overwrite existing sample metadata, convert columns to factors, and optionally specify a palette for metadata visualization.
}
\details{
This function updates the `sampleMetadata` slot of the `sampleData` component in the CYTdata object with the provided `sampleMetadata_df`.
All columns in the `sampleMetadata_df` are converted to factors (if they are not already).
If the `checkOverwrite` parameter is `TRUE` and the `sampleMetadata` slot is not empty, the user is prompted to confirm whether to overwrite the existing metadata.
The function also updates the `metadataPalette` slot with a default palette or an optional custom palette provided by the user.
}
\examples{
# Example: Import sample metadata with default palette and overwrite confirmation
updated_CYTdata <- importsampleMetadata(CYTdata = cyt_data, sampleMetadata_df = metadata_df)

# Example: Import sample metadata with a custom palette
custom_palette <- list(SampleType = c("Type1" = "red", "Type2" = "blue"))
updated_CYTdata <- importsampleMetadata(CYTdata = cyt_data, sampleMetadata_df = metadata_df, optionnalPalette = custom_palette)

}
\seealso{
\code{\link{checkValidity}} for validating the CYTdata object before and after importing metadata.
}
