% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_CYTdata_classes.R
\docType{class}
\name{Kinetic-class}
\alias{Kinetic-class}
\alias{Kinetic}
\title{Kinetic class}
\description{
This class stores data related to kinetic analysis, typically a list containing various components or results of kinetic modeling.
}
\section{Slots}{

\describe{
\item{\code{data}}{A list containing the results of kinetic analysis, which could include different types of kinetic data such as parameters, model fits, or simulation results.}
}}

