% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_Differential_Testing.R
\name{plotPiechart}
\alias{plotPiechart}
\title{Plot Piechart for Cell Abundance by Clustering}
\usage{
plotPiechart(
  CYTdata,
  clustering,
  clusters = NULL,
  samples = NULL,
  XaxisMetadata = "Timepoint",
  YaxisMetadata = "Individual",
  NFSValues = NULL,
  pieSize = TRUE,
  colorBar = "#000000"
)
}
\arguments{
\item{CYTdata}{An object containing cytometry data (class: CYTdata).}

\item{clustering}{The name of the clustering column used for grouping cells in the plot (string).}

\item{clusters}{A vector of clusters to include in the plot. If `NULL`, all clusters will be included (default is `NULL`).}

\item{samples}{A vector of sample names to include in the plot. If `NULL`, all samples are used (default is `NULL`).}

\item{XaxisMetadata}{The metadata variable to be used for the X-axis (string, default is "Timepoint").}

\item{YaxisMetadata}{The metadata variable to be used for the Y-axis (string, default is "Individual").}

\item{NFSValues}{A named vector of numeric values for scaling cell counts (optional, default is `NULL`).}

\item{pieSize}{A logical value indicating whether the size of the pie chart slices should be proportional to the total value (default is `TRUE`).}

\item{colorBar}{A string representing the color for the border of the pie chart slices (hexadecimal color, default is "#000000").}
}
\value{
A ggplot2 object representing the pie chart.
}
\description{
Generates a pie chart to visualize the cell abundance across different clusters, samples, and metadata categories.
The pie chart is stratified based on metadata values for X-axis and Y-axis, and the data can be optionally weighted
by a set of custom NFS values.
}
\details{
The pie chart is constructed by first calculating the sum of cell counts for each combination of X-axis and Y-axis metadata values.
The chart is faceted by the levels of the X-axis and Y-axis metadata, with each slice representing the abundance of cells in a given group.
If `NFSValues` is provided, the cell counts are weighted by these values. The color of the slices is determined by the clustering of cells.

If `pieSize` is set to `TRUE`, the size of each pie slice will be proportional to the sum of the cell counts for the group.
}
\examples{
# Example of creating a pie chart
piechart_plot <- plotPiechart(
  CYTdata = CYTdata,
  clustering = "Cluster1",
  samples = c("Sample1", "Sample2"),
  XaxisMetadata = "Timepoint",
  YaxisMetadata = "Individual",
  pieSize = TRUE,
  colorBar = "#FF5733"
)
print(piechart_plot)

}
\seealso{
\code{\link{CYTdata}}, \code{\link{ggplot2}}
}
