% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_Differential_Testing.R
\name{plotVolcano}
\alias{plotVolcano}
\title{Volcano Plot for Differential Testing Results}
\usage{
plotVolcano(
  CYTdata,
  differentialTestingTitle,
  thPval = 0.05,
  thFC = 2,
  displayAdjust = FALSE,
  errorNA = FALSE,
  phenoColor = TRUE,
  signifOnly = TRUE
)
}
\arguments{
\item{CYTdata}{An object containing the results of differential testing (class: CYTdata).}

\item{differentialTestingTitle}{The title of the differential testing result to visualize (string).}

\item{thPval}{The p-value threshold for significance (numeric, default is 0.05).}

\item{thFC}{The fold change threshold for significance (numeric, default is 2).}

\item{displayAdjust}{A logical value indicating whether adjusted p-values should be displayed (default is FALSE).}

\item{errorNA}{A logical value specifying whether to stop execution when there are missing adjusted p-values (default is FALSE).}

\item{phenoColor}{A logical value indicating whether the points should be color-coded by phenotype clusters (default is TRUE).}

\item{signifOnly}{A logical value indicating whether to only plot significant points (default is TRUE).}
}
\value{
A ggplot2 object representing the volcano plot.
}
\description{
Generates a volcano plot based on the differential testing results from `CYTdata`. The plot visualizes the log2 fold change (log2FC)
between two groups (`g1` and `g2`) and the significance based on p-values or adjusted p-values. The points are color-coded based on
significance or phenotype clusters.
}
\details{
The function will filter the data based on the given title (`differentialTestingTitle`). It computes the log-transformed fold
change (log2FC) and p-values (`log10Pval`). Points in the volcano plot are color-coded based on significance thresholds
for p-value and fold change.

The plot can show all points, but when `signifOnly` is set to TRUE, only the significant points are shown (colored
based on whether they are up-regulated or down-regulated).

The plot also includes dashed lines for the thresholds on the x-axis (log2FC) and y-axis (-log10(p-value)).
}
\examples{
# Example of creating a volcano plot
volcano_plot <- plotVolcano(
  CYTdata = CYTdata,
  differentialTestingTitle = "MyDifferentialTestingResult",
  thPval = 0.05,
  thFC = 2,
  displayAdjust = TRUE,
  phenoColor = TRUE,
  signifOnly = TRUE
)
print(volcano_plot)

}
\seealso{
\code{\link{CYTdata}}, \code{\link{ggplot2}}, \code{\link{ggrepel}}
}
