% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_Samples.R
\name{reorderSamples}
\alias{reorderSamples}
\title{Reorder Samples in CYTdata Object}
\usage{
reorderSamples(CYTdata, alphabetic = TRUE, newOrder = NULL)
}
\arguments{
\item{CYTdata}{A CYTdata object containing the cytometry data. The object must be valid before and after reordering the samples.}

\item{alphabetic}{A logical value indicating whether the samples should be reordered alphabetically. If `TRUE`, the samples will be sorted alphabetically by `SampleID`.
If `FALSE`, the user must provide a custom ordering in `newOrder`.}

\item{newOrder}{A character vector specifying the custom order of the samples. This argument is required if `alphabetic` is `FALSE`.}
}
\value{
A modified CYTdata object with reordered samples. The sample order will be reflected in both the `SampleID` column of `cellSample` and the rows of `sampleMetadata`.
}
\description{
This function allows users to reorder the samples in a CYTdata object either alphabetically or based on a user-provided order.
The reordering will apply to both the sample identifiers (`SampleID`) and the associated metadata in the `sampleMetadata` slot.
}
\details{
This function allows users to control the order of samples in the `CYTdata` object. If `alphabetic` is set to `TRUE`, the samples are sorted in lexicographical order
based on their `SampleID` values. If `alphabetic` is set to `FALSE`, the function expects the `newOrder` argument to provide a custom order for the samples.

The function also updates the `sampleMetadata` to reflect the new sample order.

If `newOrder` is provided but does not match the available samples, the function will raise an error.
}
\examples{
# Example: Reorder samples alphabetically
reordered_data <- reorderSamples(CYTdata = cyt_data, alphabetic = TRUE)

# Example: Reorder samples based on a custom order
custom_order <- c("SampleB", "SampleA", "SampleC")
reordered_data <- reorderSamples(CYTdata = cyt_data, alphabetic = FALSE, newOrder = custom_order)

}
\seealso{
\code{\link{checkValidity}} for validating the CYTdata object before and after reordering.
}
