% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_Batch.R
\name{runcyCombine}
\alias{runcyCombine}
\title{Perform Batch Correction on CYTdata Object}
\usage{
runcyCombine(
  CYTdata,
  batchMetadata,
  covarMetadata = NULL,
  samples = NULL,
  clustering = NULL,
  clusters = NULL,
  markers = NULL,
  normMethod = c("scale", "rank", "qnorm"),
  tiesMethod = c("average", "first", "last", "random", "max", "min"),
  seed = 92,
  xdim,
  ydim,
  rlen,
  parametric = TRUE
)
}
\arguments{
\item{CYTdata}{A CYTdata object containing the data to be batch-corrected.}

\item{batchMetadata}{A string specifying the name of the column in `sampleMetadata` for batch information.}

\item{covarMetadata}{An optional string specifying the name of the column in `sampleMetadata` for covariate information.}

\item{samples}{A character vector of sample identifiers to subset the data for batch correction.}

\item{clustering}{A string specifying the name of the column containing the clustering information (optional).}

\item{clusters}{A vector of clusters to consider for batch correction (optional).}

\item{markers}{A character vector of marker names for which the batch correction is applied.}

\item{normMethod}{A string specifying the normalization method. Options are `"scale"`, `"rank"`, and `"qnorm"`. Default is `"scale"`.}

\item{tiesMethod}{A string specifying the method for handling tied ranks. Options are `"average"`, `"first"`, `"last"`, `"random"`, `"max"`, and `"min"`. Default is `"average"`.}

\item{seed}{A numeric value for the random seed, to ensure reproducibility of the batch correction process. Default is `92`.}

\item{xdim}{The number of dimensions for the output (x-axis).}

\item{ydim}{The number of dimensions for the output (y-axis).}

\item{rlen}{The number of iterations for the batch correction.}

\item{parametric}{A logical indicating whether to use parametric methods for batch correction. Default is `TRUE`.}
}
\value{
A CYTdata object with the batch-corrected data.
}
\description{
This function applies batch correction to a `CYTdata` object, using the specified batch and covariate metadata. The batch correction is performed by utilizing the `cyCombine::batch_correct` function, and it can handle multiple types of normalizations and tie-breaking methods. The function also allows for the inclusion of optional clustering and marker information.
}
\details{
This function performs batch correction on the `CYTdata` object using the specified batch and optional covariate metadata. The correction method can be customized using the `normMethod` and `tiesMethod` parameters. The function also supports subsetting the data to specific samples or clusters, and the corrected expression data is returned in the CYTdata object. The `cyCombine::batch_correct` function is used for the core batch correction process.
}
\examples{
# Example 1: Perform batch correction using the "batch" metadata
CYTdata_corrected <- runcyCombine(CYTdata = CYTdata, batchMetadata = "batch", markers = c("marker1", "marker2"))

# Example 2: Perform batch correction with parametric method and custom seed
CYTdata_corrected <- runcyCombine(CYTdata = CYTdata, batchMetadata = "batch", markers = c("marker1", "marker2"), seed = 100)

# Example 3: Perform batch correction for specific samples and clusters
CYTdata_corrected <- runcyCombine(CYTdata = CYTdata, batchMetadata = "batch", samples = c("sample1", "sample2"), clustering = "cluster", clusters = c("cluster1", "cluster2"))

}
\seealso{
\code{\link{batch_correct}} from the \pkg{cyCombine} package for batch correction functionality.
}
