% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_DimRed.R
\name{runDimRed}
\alias{runDimRed}
\title{Run Dimensionality Reduction}
\usage{
runDimRed(
  CYTdata,
  name,
  markers = NULL,
  type = c("UMAP", "tSNE", "lvish", "PCA"),
  seed = 42,
  checkOverwrite = TRUE,
  ...
)
}
\arguments{
\item{CYTdata}{An object of class \code{CYTdata}. It contains the cytometry data and related metadata.}

\item{name}{A character string specifying the name of the new dimensionality reduction result. This will be used to store the result in the \code{cellDimRed} slot.}

\item{markers}{A character vector specifying the marker names to be used for dimensionality reduction. If \code{NULL}, all available markers in the data will be used.}

\item{type}{A character string specifying the type of dimensionality reduction method to be applied. It can be one of: \code{"UMAP"}, \code{"tSNE"}, \code{"lvish"}, or \code{"PCA"}.}

\item{seed}{An integer value to set the seed for random number generation. Default is \code{42}.}

\item{checkOverwrite}{A logical value indicating whether to check for the existence of the specified \code{name} in the \code{cellDimRed} slot and prompt the user for confirmation before overwriting. Default is \code{TRUE}.}

\item{\dots}{Additional parameters passed to the respective dimensionality reduction functions (e.g., \code{uwot::umap}, \code{Rtsne::Rtsne}, or \code{stats::prcomp}).}
}
\value{
The updated \code{CYTdata} object, with the new dimensionality reduction result stored in the \code{cellDimRed} slot.
}
\description{
This function performs dimensionality reduction on the provided cytometry data using one of the selected methods (UMAP, tSNE, LVish, or PCA). The function stores the results in the `cellDimRed` slot of the `CYTdata` object. It allows for the application of different dimensionality reduction algorithms to explore and visualize high-dimensional cytometry data in a lower-dimensional space (typically 2D or 3D).
}
\details{
This function supports the following dimensionality reduction methods:
- \code{UMAP}: Uniform Manifold Approximation and Projection.
- \code{tSNE}: t-Distributed Stochastic Neighbor Embedding.
- \code{lvish}: LVish manifold learning method (a non-linear method).
- \code{PCA}: Principal Component Analysis.

The results of the selected dimensionality reduction method are stored in the \code{cellDimRed} slot of the \code{CYTdata} object under the specified \code{name}. This can then be used for visualization or further analysis.
}
\examples{
# Example usage:
CYTdata <- runDimRed(CYTdata = CYTdata,
                     name = "UMAP_result",
                     markers = c("CD3", "CD8", "CD45"),
                     type = "UMAP")

# Example with PCA
CYTdata <- runDimRed(CYTdata = CYTdata,
                     name = "PCA_result",
                     markers = c("CD3", "CD4", "CD56"),
                     type = "PCA",
                     seed = 123)

}
