% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_CYTdata_classes.R
\docType{class}
\name{sampleData-class}
\alias{sampleData-class}
\alias{sampleData}
\title{sampleData class}
\description{
This class stores sample-related data, including sample-specific metadata and any additional sample-related data.
}
\section{Slots}{

\describe{
\item{\code{cellSample}}{A data frame containing cell-level data for the sample.}

\item{\code{sampleMetadata}}{A data frame containing metadata for each sample.}

\item{\code{metadataPalette}}{A list containing color palettes used for the sample metadata visualization.}

\item{\code{sampleAdditionaldata}}{A data frame containing any additional data associated with the sample.}
}}

