% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_Palettes.R
\name{showPalettes}
\alias{showPalettes}
\title{Show Color Palettes for Metadata or Clustering}
\usage{
showPalettes(
  CYTdata,
  type = c("metadata", "clustering"),
  metadata = NULL,
  clustering = NULL,
  ncol = 10,
  labelSize = 5
)
}
\arguments{
\item{CYTdata}{An object of class \code{CYTdata} containing cytometry data and associated metadata or clustering information.}

\item{type}{A character string specifying whether to show the palette for "metadata" or "clustering". Must be one of "metadata" or "clustering".}

\item{metadata}{A character string specifying the metadata column name (if \code{type = "metadata"}). Ignored if \code{type = "clustering"}.}

\item{clustering}{A character string specifying the clustering column name (if \code{type = "clustering"}). Ignored if \code{type = "metadata"}.}

\item{ncol}{An integer indicating the number of columns to display the palette in. Default is 10.}

\item{labelSize}{A numeric value specifying the size of the labels displayed in the palette plot. Default is 5.}
}
\value{
A \code{ggplot} object displaying the color palette for the chosen metadata or clustering.
}
\description{
This function visualizes the color palette associated with either metadata or clustering information in the given `CYTdata` object. It allows users to view the color assignments for categories in metadata or clustering results and provides a plot to show the color distribution.
}
\details{
This function extracts the color palette either from the metadata or the clustering information stored in the `CYTdata` object. It then arranges the colors in a grid format and labels each color with the corresponding category name. The plot shows the colors with the labels, allowing the user to see how each category is represented by color.

The `ncol` parameter controls the number of columns in the palette plot, and `labelSize` adjusts the font size of the category labels on the plot.
}
\examples{
# Example usage:
# Display the color palette for a metadata column
showPalettes(CYTdata, type = "metadata", metadata = "cellType", ncol = 5, labelSize = 6)

# Display the color palette for a clustering result
showPalettes(CYTdata, type = "clustering", clustering = "cluster1", ncol = 8, labelSize = 5)

}
