% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_CYTdata_classes.R
\docType{class}
\name{TrajectoryInference-class}
\alias{TrajectoryInference-class}
\alias{TrajectoryInference}
\title{TrajectoryInference class}
\description{
This class is used to store information about a trajectory inference analysis, including details about the root and leaf cells,
the inferred network, walk paths, and differential trajectory results.
}
\section{Slots}{

\describe{
\item{\code{root.cells}}{A vector containing the identifiers of the root cells in the trajectory analysis.}

\item{\code{leaf.cells}}{A vector containing the identifiers of the leaf cells in the trajectory analysis.}

\item{\code{network}}{A list containing the network representation of the trajectory inference.}

\item{\code{walk}}{A list that stores the inferred walk paths through the trajectory network.}

\item{\code{diff.traj}}{A list containing the results of differential trajectory analysis.}
}}

